Fix Text Tool Overwriting Existing Text on Editing Text Layer (#1356)
fix: text edit overwriting existing text

+++ editor/src/messages/tool/tool_messages/text_tool.rs
@@ -330,7 +330,7 @@ impl TextToolData {
 		else if let Some(editing_text) = self.editing_text.as_ref().filter(|_| state == TextToolFsmState::Ready) {
 			responses.add(DocumentMessage::StartTransaction);
 
-			let network = new_text_network(String::new(), editing_text.font.clone(), editing_text.font_size as f32);
+			let network = new_text_network(String::new(), editing_text.font.clone(), editing_text.font_size);
 
 			responses.add(Operation::AddFrame {
 				path: self.layer_path.clone(),
